Saving Fixed and Code More Readable
Also Changed colors
